Implement/Use RoutesToSwaggerDocs::Base
